Add shared titleInfoNode to UTM module
