+++ client/my-sites/stats/features/modules/stats-utm/stats-module-utm.jsx
@@ -1,16 +1,13 @@
 import config from '@automattic/calypso-config';
 import { StatsCard } from '@automattic/components';
-import { localizeUrl } from '@automattic/i18n-utils';
 import { trendingUp } from '@wordpress/icons';
 import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
 import { useState } from 'react';
-import StatsInfoArea from 'calypso/my-sites/stats/features/modules/shared/stats-info-area';
 import { useSelector } from 'calypso/state';
-import { getSiteSlug, isJetpackSite } from 'calypso/state/sites/selectors';
+import { getSiteSlug } from 'calypso/state/sites/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import EmptyModuleCard from '../../../components/empty-module-card/empty-module-card';
-import { JETPACK_SUPPORT_URL_TRAFFIC, SUPPORT_URL } from '../../../const';
 import useUTMMetricsQuery from '../../../hooks/use-utm-metrics-query';
 import ErrorPanel from '../../../stats-error';
 import StatsListCard from '../../../stats-list/stats-list-card';
@@ -20,6 +17,8 @@ import { StatsEmptyActionUTMBuilder } from '../shared';
 import StatsCardSkeleton from '../shared/stats-card-skeleton';
 import UTMDropdown from './stats-module-utm-dropdown';
 import UTMExportButton from './utm-export-button';
+import useUTMSupportURL from './utm-support-url';
+import UTMTitleInfoNode from './utm-title-info-node';
 
 import '../../../stats-module/style.scss';
 import '../../../stats-list/style.scss';
@@ -129,28 +128,8 @@ const StatsModuleUTM = ( {
 		? generateFileNameForDownload( siteSlug, period )
 		: '';
 
-	const isSiteJetpackNotAtomic = useSelector( ( state ) =>
-		isJetpackSite( state, siteId, { treatAtomicAsJetpackSite: false } )
-	);
-
-	const supportUrl = isSiteJetpackNotAtomic
-		? localizeUrl( `${ JETPACK_SUPPORT_URL_TRAFFIC }#harnessing-utm-stats-for-precision-tracking` )
-		: localizeUrl( `${ SUPPORT_URL }#utm` );
-
-	const titleNodes = (
-		<StatsInfoArea isNew>
-			{ translate(
-				'Track your campaign {{link}}UTM performance data{{/link}}. Generate URL codes with our builder.',
-				{
-					comment: '{{link}} links to support documentation.',
-					components: {
-						link: <a target="_blank" rel="noreferrer" href={ supportUrl } />,
-					},
-					context: 'Stats: Popover information when the UTM module has data',
-				}
-			) }
-		</StatsInfoArea>
-	);
+	const supportUrl = useUTMSupportURL( siteId );
+	const titleInfoNode = <UTMTitleInfoNode supportUrl={ supportUrl } />;
 
 	return (
 		<>
@@ -169,7 +148,7 @@ const StatsModuleUTM = ( {
 							<StatsCard
 								className={ clsx( 'stats-card--empty-variant', className ) }
 								title={ moduleStrings.title }
-								titleNodes={ <StatsInfoArea isNew /> }
+								titleNodes={ titleInfoNode }
 								isEmpty
 								emptyMessage={
 									<EmptyModuleCard
@@ -206,7 +185,7 @@ const StatsModuleUTM = ( {
 									data={ data }
 									useShortLabel={ useShortLabel }
 									title={ moduleStrings?.title }
-									titleNodes={ titleNodes }
+									titleNodes={ titleInfoNode }
 									emptyMessage={ <div>{ moduleStrings.empty }</div> }
 									metricLabel={ metricLabel }
 									showMore={
@@ -250,6 +229,7 @@ const StatsModuleUTM = ( {
 						) }
 				</>
 			) }
+			{ /* TODO: Refactor later */ }
 			{ ! isNewEmptyStateEnabled && (
 				<>
 					<StatsListCard
